Split user and password states
