+++ src/services/user.js
@@ -4,8 +4,6 @@ import {
     postAuthLogin, 
     getAuthUser, 
     patchAuthUser, 
-    postPasswordReset, 
-    postPasswordUpdate,
     postAuthLogout,
 } from "./api";
 
@@ -103,38 +101,6 @@ export function logoutUser() {
     };
 }
 
-export function resetPassword(email) {
-    return function (dispatch) {
-        dispatch(passwordResetRequest());
-        postPasswordReset(email).then(response => {
-            if (response && response.success) {
-                dispatch(passwordResetSuccess());
-            } else {
-                dispatch(passwordResetError());
-            }
-        }).catch(e => {
-            console.log(`Exception occurred while password reset ${e}`);
-            dispatch(passwordResetError());
-        });
-    };
-}
-
-export function updatePassword(password, code) {
-    return function (dispatch) {
-        dispatch(passwordUpdateRequest());
-        postPasswordUpdate(password, code).then(response => {
-            if (response && response.success) {
-                dispatch(passwordUpdateSuccess());
-            } else {
-                dispatch(passwordUpdateError());
-            }
-        }).catch(e => {
-            console.log(`Exception occurred while password reset ${e}`);
-            dispatch(passwordUpdateError());
-        });
-    };
-}
-
 const initialState = {
     user: null,
     userChecked: false,
@@ -144,12 +110,6 @@ const initialState = {
     loginError: false,
     logoutRequest: false,
     logoutError: false,
-    passwordResetRequest: false,
-    passwordResetError: false,
-    passwordResetSuccess: false,
-    passwordUpdateRequest: false,
-    passwordUpdateError: false,
-    passwordUpdateSuccess: false,
 };
 
 const slice = createSlice({
@@ -191,36 +151,6 @@ const slice = createSlice({
             state.logoutRequest = false;
             state.logoutError = true;
         },
-        passwordResetRequest(state, action) {
-            state.passwordResetRequest = true;
-            state.passwordResetError = false;
-            state.passwordResetSuccess = false;
-        },
-        passwordResetError(state, action) {
-            state.passwordResetRequest = false;
-            state.passwordResetError = true;
-            state.passwordResetSuccess = false;
-        },
-        passwordResetSuccess(state, action) {
-            state.passwordResetRequest = false;
-            state.passwordResetError = false;
-            state.passwordResetSuccess = true;
-        },
-        passwordUpdateRequest(state, action) {
-            state.passwordUpdateRequest = true;
-            state.passwordUpdateError = false;
-            state.passwordUpdateSuccess = false;
-        },
-        passwordUpdateError(state, action) {
-            state.passwordUpdateRequest = false;
-            state.passwordUpdateError = true;
-            state.passwordUpdateSuccess = false;
-        },
-        passwordUpdateSuccess(state, action) {
-            state.passwordUpdateRequest = false;
-            state.passwordUpdateError = false;
-            state.passwordUpdateSuccess = true;
-        },
     }
 });
 
@@ -232,12 +162,6 @@ export const {
     loginRequestError,
     logoutRequest,
     logoutRequestError,
-    passwordResetRequest,
-    passwordResetError,
-    passwordResetSuccess,
-    passwordUpdateRequest,
-    passwordUpdateError,
-    passwordUpdateSuccess,
 } = slice.actions;
 
 export default slice;
